+++ src/renderable/trigger.js
@@ -1,7 +1,6 @@
 import Renderable from "./renderable.js";
 import collision from "./../physics/collision.js";
 import Body from "./../physics/body.js";
-import Rect from "./../geometries/rectangle.js";
 import level from "./../level/level.js";
 import { world, viewport } from "./../game.js";
 
@@ -66,9 +65,16 @@ class Trigger extends Renderable {
             }
         }.bind(this));
 
-
-        // physic body to check for collision against
-        this.body = new Body(this, settings.shapes || new Rect(0, 0, this.width, this.height));
+        // add and configure the physic body
+        var shape = settings.shapes;
+        if (typeof shape === "undefined") {
+            shape = pool.pull("Polygon", 0, 0, [
+                pool.pull("Vector2d", 0,          0),
+                pool.pull("Vector2d", this.width, 0),
+                pool.pull("Vector2d", this.width, this.height)
+            ]);
+        }
+        this.body = new Body(this, shape);
         this.body.collisionType = collision.types.ACTION_OBJECT;
         // by default only collides with PLAYER_OBJECT
         this.body.setCollisionMask(collision.types.PLAYER_OBJECT);